requirements.txt
pyspark>=3.5.0
apache-airflow>=2.0.0
pandas
numpy
faker
psycopg2-binary
pyyaml